+++ docs/.vitepress/theme/index.ts
@@ -1,0 +1,11 @@
+import type { Theme } from 'vitepress'
+import DefaultTheme from 'vitepress/theme'
+import { h } from 'vue'
+import LegalFooter from './LegalFooter.vue'
+
+export default {
+  extends: DefaultTheme,
+  Layout: () => h(DefaultTheme.Layout, null, {
+    'layout-bottom': () => h(LegalFooter),
+  }),
+} satisfies Theme